pyproject.toml
[project]
name = "sbsp-dispatch"
version = "0.1.0"
description = ""
authors = [
    {name = "Lars Schellhas",email = "lars@schellhas.engineering"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pyomo (>=6.9.2,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "entsoe-py (>=0.6.18,<0.7.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "highspy (>=1.10.0,<2.0.0)",
    "plotly (>=6.0.1,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
